+++ pom/pages/HomePage.js
@@ -1,5 +1,5 @@
 import { step } from 'allure-js-commons';
-import HeaderComponent from '../components/HeaderComponent';
+import HeaderComponent from '../components/Header';
 
 export default class HomePage {
     constructor(page) {
